Backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
apscheduler
h5py
pandas
numpy
scipy
matplotlib